presplit datasets, added random insertion
